requirements.txt
numpy
pillow
tqdm
pyyaml
scikit-learn

torch
torchvision
timm
segmentation-models-pytorch

